+++ tests/nativeapp/test_manager.py
@@ -151,7 +151,7 @@ def test_sync_deploy_root_with_stage(
         role="new_role",
         prune=True,
         recursive=True,
-        stage_path=DefaultStagePathParts.from_fqn(stage_fqn),
+        stage_path_parts=DefaultStagePathParts.from_fqn(stage_fqn),
     )
 
     mock_stage_exists.assert_called_once_with(stage_fqn)
@@ -167,6 +167,7 @@ def test_sync_deploy_root_with_stage(
         deploy_root_path=dm.project_root / pkg_model.deploy_root,
         diff_result=mock_diff_result,
         stage_full_path="app_pkg.app_src.stage",
+        force_overwrite=False,
     )
 
 
@@ -215,7 +216,7 @@ def test_sync_deploy_root_with_stage_subdir(
         role="new_role",
         prune=True,
         recursive=True,
-        stage_path=DefaultStagePathParts.from_fqn(stage_fqn, "v1"),
+        stage_path_parts=DefaultStagePathParts.from_fqn(stage_fqn, "v1"),
     )
 
     mock_stage_exists.assert_called_once_with(stage_fqn)
@@ -231,6 +232,7 @@ def test_sync_deploy_root_with_stage_subdir(
         deploy_root_path=dm.project_root / pkg_model.deploy_root,
         diff_result=mock_diff_result,
         stage_full_path=stage_full_path,
+        force_overwrite=False,
     )
 
 
@@ -285,7 +287,7 @@ def test_sync_deploy_root_with_stage_prune(
         role="new_role",
         prune=prune,
         recursive=True,
-        stage_path=DefaultStagePathParts.from_fqn(stage_fqn),
+        stage_path_parts=DefaultStagePathParts.from_fqn(stage_fqn),
     )
 
     if expected_warn: